fix libft

+++ libft/srcs/printf/converters/char.c
@@ -65,7 +65,7 @@ void	handle_null_str(t_arg *arg)
 	len = 6;
 	if (arg->precision > -1 && len > (size_t)arg->precision)
 		len = arg->precision;
-	if ((new = ft_strnew(((size_t)arg->width < len) ? arg->width : len)))
+	if ((new = ft_strnew(((size_t)arg->width < len) ? (size_t)arg->width : len)))
 	{
 		while (!arg->left && i + len < (size_t)arg->width)
 			new[i++] = arg->zero ? '0' : ' ';
